refactor: Move ScoreAnalyze to components
