+++ src/components/ScoreAnalyze/scoreAnalyze.module.scss
